requirements.txt
# AIMET + Power-of-2 Symmetric Quantization Requirements

# PyTorch ecosystem (AIMET compatible versions)
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# Core AIMET packages (latest stable)
aimet-torch==2.11.0

# Data handling and utilities
numpy>=1.24.0
Pillow>=10.0.0

# Configuration and output
pyyaml>=6.0
tqdm>=4.65.0

# Scientific computing
scipy>=1.13.0

# Optional utilities
matplotlib>=3.8.0
seaborn>=0.13.0

# Development and linting tools
pytest>=7.4.0
flake8>=6.0.0
